requirements.txt
flask>=2.3.0
opencv-python>=4.8.0
mediapipe==0.10.14
numpy>=1.24.0,<2.0.0
pyttsx3>=2.90
requests>=2.31.0
python-dotenv>=1.0.0
protobuf==4.25.3
attrs>=19.1.0
